requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
passlib[argon2]
python-multipart
stripe
fpdf2
arabic-reshaper
python-bidi
sentry-sdk[fastapi]
python-dotenv
pytest
httpx
